Move HTML Builder back into Formatter namespace
