+++ lib/cucumber/formatter/cucumber.css
