+++ lib/cucumber/formatter/inline-js.js
