+++ lib/cucumber/formatter/jquery-min.js
